fix(agent): CI 补丁 —— sentTokens 记账 Int→Long 显式转换
TokenEstimator.estimate 返回 Int，全量记账分支直接赋给 Long 型
sentTokens 触发类型不匹配（AgentRunner.kt:400，CI run 35973734052）。
增量分支的 += 本就隐式加宽，无需改动。

+++ core-agent/src/main/java/com/rickeal/agent/core/agent/AgentRunner.kt
@@ -397,7 +397,7 @@ class AgentRunner(
                 val requestMessages = generationRequest.messages
                 if (contextVersion != accountedVersion || request.conversationId != lastCid) {
                     accountedIds = requestMessages.map { it.id }.toMutableSet()
-                    sentTokens = TokenEstimator.estimate(requestMessages)
+                    sentTokens = TokenEstimator.estimate(requestMessages).toLong()
                     accountedVersion = contextVersion
                 } else {
                     val fresh = requestMessages.filter { it.id !in accountedIds }
